Cleanup: AST: 5: Merge position.rs into location.rs
Move SourceLocation from position.rs into location.rs to keep tightly
coupled data structures together. Update all imports and delete position.rs.
