+++ src/txxt/ast/location.rs
@@ -1,9 +1,11 @@
 //! Position and location tracking for source code locations
 //!
 //! This module defines the data structures for representing positions
-//! and locations in source code.
+//! and locations in source code, as well as utilities for converting
+//! byte offsets to line/column positions.
 
 use std::fmt;
+use std::ops::Range;
 
 /// Represents a position in source code (line and column)
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -56,6 +58,57 @@ impl Location {
 impl fmt::Display for Location {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         write!(f, "{}..{}", self.start, self.end)
+    }
+}
+
+/// Provides fast conversion from byte offsets to line/column positions
+pub struct SourceLocation {
+    /// Byte offsets where each line starts
+    line_starts: Vec<usize>,
+}
+
+impl SourceLocation {
+    /// Create a new SourceLocation from source code
+    pub fn new(source: &str) -> Self {
+        let mut line_starts = vec![0];
+
+        for (byte_pos, ch) in source.char_indices() {
+            if ch == '\n' {
+                line_starts.push(byte_pos + 1);
+            }
+        }
+
+        Self { line_starts }
+    }
+
+    /// Convert a byte offset to a line/column position
+    pub fn byte_to_position(&self, byte_offset: usize) -> Position {
+        let line = self
+            .line_starts
+            .binary_search(&byte_offset)
+            .unwrap_or_else(|i| i - 1);
+
+        let column = byte_offset - self.line_starts[line];
+
+        Position::new(line, column)
+    }
+
+    /// Convert a byte range to a location
+    pub fn range_to_location(&self, range: &Range<usize>) -> Location {
+        Location::new(
+            self.byte_to_position(range.start),
+            self.byte_to_position(range.end),
+        )
+    }
+
+    /// Get the total number of lines in the source
+    pub fn line_count(&self) -> usize {
+        self.line_starts.len()
+    }
+
+    /// Get the byte offset for the start of a line
+    pub fn line_start(&self, line: usize) -> Option<usize> {
+        self.line_starts.get(line).copied()
     }
 }
 
@@ -144,5 +197,73 @@ mod tests {
     fn test_location_display() {
         let location = Location::new(Position::new(1, 0), Position::new(2, 5));
         assert_eq!(format!("{}", location), "1:0..2:5");
+    }
+
+    #[test]
+    fn test_byte_to_position_single_line() {
+        let loc = SourceLocation::new("Hello");
+        assert_eq!(loc.byte_to_position(0), Position::new(0, 0));
+        assert_eq!(loc.byte_to_position(1), Position::new(0, 1));
+        assert_eq!(loc.byte_to_position(4), Position::new(0, 4));
+    }
+
+    #[test]
+    fn test_byte_to_position_multiline() {
+        let loc = SourceLocation::new("Hello\nworld\ntest");
+
+        // First line
+        assert_eq!(loc.byte_to_position(0), Position::new(0, 0));
+        assert_eq!(loc.byte_to_position(5), Position::new(0, 5));
+
+        // Second line
+        assert_eq!(loc.byte_to_position(6), Position::new(1, 0));
+        assert_eq!(loc.byte_to_position(10), Position::new(1, 4));
+
+        // Third line
+        assert_eq!(loc.byte_to_position(12), Position::new(2, 0));
+        assert_eq!(loc.byte_to_position(15), Position::new(2, 3));
+    }
+
+    #[test]
+    fn test_byte_to_position_with_unicode() {
+        let loc = SourceLocation::new("Hello\nwörld");
+        // Unicode characters take multiple bytes
+        assert_eq!(loc.byte_to_position(6), Position::new(1, 0));
+        assert_eq!(loc.byte_to_position(7), Position::new(1, 1));
+    }
+
+    #[test]
+    fn test_range_to_location_single_line() {
+        let loc = SourceLocation::new("Hello World");
+        let location = loc.range_to_location(&(0..5));
+
+        assert_eq!(location.start, Position::new(0, 0));
+        assert_eq!(location.end, Position::new(0, 5));
+    }
+
+    #[test]
+    fn test_range_to_location_multiline() {
+        let loc = SourceLocation::new("Hello\nWorld\nTest");
+        let location = loc.range_to_location(&(6..12));
+
+        assert_eq!(location.start, Position::new(1, 0));
+        assert_eq!(location.end, Position::new(2, 0));
+    }
+
+    #[test]
+    fn test_line_count() {
+        assert_eq!(SourceLocation::new("single").line_count(), 1);
+        assert_eq!(SourceLocation::new("line1\nline2").line_count(), 2);
+        assert_eq!(SourceLocation::new("line1\nline2\nline3").line_count(), 3);
+    }
+
+    #[test]
+    fn test_line_start() {
+        let loc = SourceLocation::new("Hello\nWorld\nTest");
+
+        assert_eq!(loc.line_start(0), Some(0));
+        assert_eq!(loc.line_start(1), Some(6));
+        assert_eq!(loc.line_start(2), Some(12));
+        assert_eq!(loc.line_start(3), None);
     }
 }
